requirements.txt
autopep8==1.5.7
configobj==5.0.6
pprintpp==0.4.0
prettyprint==0.1.5
pycodestyle==2.7.0
python-dateutil==2.8.1
PyYAML==5.4.1
ruamel.std.argparse==0.8.3
ruamel.yaml.clib==0.2.2
ruamel.yaml.convert==0.3.2
six==1.16.0
toml==0.10.2
